+++ src/main/java/carracing/domain/RaceResult.java
@@ -1,4 +1,4 @@
-package carracing.util;
+package carracing.domain;
 
 /*
  * CarRacing
