+++ src/main/java/carracing/domain/Winners.java
@@ -1,4 +1,4 @@
-package carracing.util;
+package carracing.domain;
 
 import java.util.List;
 import java.util.stream.Collectors;
