+++ src/test/java/carracing/CarTest.java
@@ -1,7 +1,7 @@
 package carracing;
 
+import carracing.domain.Car;
 import carracing.exception.CarNameFormatException;
-import carracing.util.Car;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
